add active job to send sms messages
