+++ app/admin/messages.rb
@@ -9,20 +9,8 @@ ActiveAdmin.register Message do
 
       respond_to do |format|
         if @message.save
-          begin
-            init_twilio
-            sms = @client.messages.create(
-                from: ENV['TWILIO_NUMBER'],
-                to: @user.contact_number,
-                body: @message.message
-            )
-          rescue Twilio::REST::RestError => e
-            format.html { redirect_to "/admin/users/#{@user.id}/create_sms", notice: "Error: unable to send message to " + @user.name + ", invalid contact number" }
-          rescue StandardError => e
-            format.html { redirect_to "/admin/users/#{@user.id}/create_sms", notice: "Error: unable to send message to " + @user.name }
-          else
-            format.html { redirect_to admin_users_path, notice: "Message was successfully sent to #{@user.name}." }
-          end
+          SendTextMessageJob.perform_later(@user.contact_number, @message.message)
+          format.html { redirect_to admin_users_path, notice: "Message was successfully sent to #{@user.name}." }
         else
           format.html { redirect_to "/admin/users/#{@user.id}/create_sms", notice: "Error: unable to send message to " + @user.name }
         end
@@ -31,12 +19,6 @@ ActiveAdmin.register Message do
 
     def message_params
       params.require(:message).permit(:message, :user_id)
-    end
-
-    def init_twilio
-      account_sid = ENV['TWILIO_SID']
-      auth_token = ENV['TWILIO_TOKEN']
-      @client = Twilio::REST::Client.new account_sid, auth_token
     end
   end
 end
